Crate rusoto_lambda

AWS Lambda
Overview

This is the AWS Lambda API Reference. The AWS Lambda Developer Guide provides additional information. For the service overview, see What is AWS Lambda, and for information about how the service works, see AWS Lambda: How it Works in the AWS Lambda Developer Guide.

If you're using the service, you're probably looking for LambdaClient and Lambda.

Structs

AccountLimit
Limits that are related to concurrency and code storage. All file and storage sizes are in bytes.

AccountUsage
The number of functions and amount of storage in use.

AddLayerVersionPermissionRequest

AddLayerVersionPermissionResponse

AddPermissionRequest

AddPermissionResponse

AliasConfiguration
Provides configuration information about a Lambda function alias.

AliasRoutingConfiguration
The traffic-shifting configuration of a Lambda function alias.

Concurrency

CreateAliasRequest

CreateEventSourceMappingRequest

CreateFunctionRequest

DeadLetterConfig
The dead letter queue for failed asynchronous invocations.

DeleteAliasRequest

DeleteEventSourceMappingRequest

DeleteFunctionConcurrencyRequest

DeleteFunctionRequest

DeleteLayerVersionRequest

Environment
A function's environment variable settings.

EnvironmentError
Error messages for environment variables that couldn't be applied.

EnvironmentResponse
The results of a configuration update that applied environment variables.

EventSourceMappingConfiguration
A mapping between an AWS resource and an AWS Lambda function. See CreateEventSourceMapping for details.

FunctionCode
The code for the Lambda function. You can specify either an object in Amazon S3, or upload a deployment package directly.

FunctionCodeLocation
Details about a function's deployment package.

FunctionConfiguration
Details about a function's configuration.

GetAccountSettingsRequest

GetAccountSettingsResponse

GetAliasRequest

GetEventSourceMappingRequest

GetFunctionConfigurationRequest

GetFunctionRequest

GetFunctionResponse

GetLayerVersionByArnRequest

GetLayerVersionPolicyRequest

GetLayerVersionPolicyResponse

GetLayerVersionRequest

GetLayerVersionResponse

GetPolicyRequest

GetPolicyResponse

InvocationRequest

InvocationResponse

InvokeAsyncRequest

InvokeAsyncResponse
A success response (202 Accepted) indicates that the request is queued for invocation.

LambdaClient
A client for the AWS Lambda API.

Layer
An AWS Lambda layer.

LayerVersionContentInput
A ZIP archive that contains the contents of an AWS Lambda layer. You can specify either an Amazon S3 location, or upload a layer archive directly.

LayerVersionContentOutput
Details about a version of an AWS Lambda layer.

LayerVersionsListItem
Details about a version of an AWS Lambda layer.

LayersListItem
Details about an AWS Lambda layer.

ListAliasesRequest

ListAliasesResponse

ListEventSourceMappingsRequest

ListEventSourceMappingsResponse

ListFunctionsRequest

ListFunctionsResponse
A list of Lambda functions.

ListLayerVersionsRequest

ListLayerVersionsResponse

ListLayersRequest

ListLayersResponse

ListTagsRequest

ListTagsResponse

ListVersionsByFunctionRequest

ListVersionsByFunctionResponse

PublishLayerVersionRequest

PublishLayerVersionResponse

PublishVersionRequest

PutFunctionConcurrencyRequest

RemoveLayerVersionPermissionRequest

RemovePermissionRequest

TagResourceRequest

TracingConfig
The function's AWS X-Ray tracing configuration.

TracingConfigResponse
The function's AWS X-Ray tracing configuration.

UntagResourceRequest

UpdateAliasRequest

UpdateEventSourceMappingRequest

UpdateFunctionCodeRequest

UpdateFunctionConfigurationRequest

VpcConfig
The VPC security groups and subnets that are attached to a Lambda function.

VpcConfigResponse
The VPC security groups and subnets that are attached to a Lambda function.

Enums

AddLayerVersionPermissionError
Errors returned by AddLayerVersionPermission

AddPermissionError
Errors returned by AddPermission

CreateAliasError
Errors returned by CreateAlias

CreateEventSourceMappingError
Errors returned by CreateEventSourceMapping

CreateFunctionError
Errors returned by CreateFunction

DeleteAliasError
Errors returned by DeleteAlias

DeleteEventSourceMappingError
Errors returned by DeleteEventSourceMapping

DeleteFunctionConcurrencyError
Errors returned by DeleteFunctionConcurrency

DeleteFunctionError
Errors returned by DeleteFunction

DeleteLayerVersionError
Errors returned by DeleteLayerVersion

GetAccountSettingsError
Errors returned by GetAccountSettings

GetAliasError
Errors returned by GetAlias

GetEventSourceMappingError
Errors returned by GetEventSourceMapping

GetFunctionConfigurationError
Errors returned by GetFunctionConfiguration

GetFunctionError
Errors returned by GetFunction

GetLayerVersionByArnError
Errors returned by GetLayerVersionByArn

GetLayerVersionError
Errors returned by GetLayerVersion

GetLayerVersionPolicyError
Errors returned by GetLayerVersionPolicy

GetPolicyError
Errors returned by GetPolicy

InvokeAsyncError
Errors returned by InvokeAsync

InvokeError
Errors returned by Invoke

ListAliasesError
Errors returned by ListAliases

ListEventSourceMappingsError
Errors returned by ListEventSourceMappings

ListFunctionsError
Errors returned by ListFunctions

ListLayerVersionsError
Errors returned by ListLayerVersions

ListLayersError
Errors returned by ListLayers

ListTagsError
Errors returned by ListTags

ListVersionsByFunctionError
Errors returned by ListVersionsByFunction

PublishLayerVersionError
Errors returned by PublishLayerVersion

PublishVersionError
Errors returned by PublishVersion

PutFunctionConcurrencyError
Errors returned by PutFunctionConcurrency

RemoveLayerVersionPermissionError
Errors returned by RemoveLayerVersionPermission

RemovePermissionError
Errors returned by RemovePermission

TagResourceError
Errors returned by TagResource

UntagResourceError
Errors returned by UntagResource

UpdateAliasError
Errors returned by UpdateAlias

UpdateEventSourceMappingError
Errors returned by UpdateEventSourceMapping

UpdateFunctionCodeError
Errors returned by UpdateFunctionCode

UpdateFunctionConfigurationError
Errors returned by UpdateFunctionConfiguration

Traits

Lambda
Trait representing the capabilities of the AWS Lambda API. AWS Lambda clients implement this trait.