Enum rusoto_lambda::InvokeError

pub enum InvokeError {
    EC2AccessDenied(String),
    EC2Throttled(String),
    EC2Unexpected(String),
    ENILimitReached(String),
    InvalidParameterValue(String),
    InvalidRequestContent(String),
    InvalidRuntime(String),
    InvalidSecurityGroupID(String),
    InvalidSubnetID(String),
    InvalidZipFile(String),
    KMSAccessDenied(String),
    KMSDisabled(String),
    KMSInvalidState(String),
    KMSNotFound(String),
    RequestTooLarge(String),
    ResourceNotFound(String),
    Service(String),
    SubnetIPAddressLimitReached(String),
    TooManyRequests(String),
    UnsupportedMediaType(String),
    HttpDispatch(HttpDispatchError),
    Credentials(CredentialsError),
    Validation(String),
    Unknown(String),
}

Errors returned by Invoke

Variants
EC2AccessDenied(String)

EC2Throttled(String)

AWS Lambda was throttled by Amazon EC2 during Lambda function initialization using the execution role provided for the Lambda function.

EC2Unexpected(String)

AWS Lambda received an unexpected EC2 client exception while setting up for the Lambda function.

ENILimitReached(String)

AWS Lambda was not able to create an Elastic Network Interface (ENI) in the VPC, specified as part of Lambda function configuration, because the limit for network interfaces has been reached.

InvalidParameterValue(String)

One of the parameters in the request is invalid. For example, if you provided an IAM role for AWS Lambda to assume in the CreateFunction or the UpdateFunctionConfiguration API, that AWS Lambda is unable to assume you will get this exception. You will also get this exception if you have selected a deprecated runtime, such as Node v0.10.42.

InvalidRequestContent(String)

The request body could not be parsed as JSON.

InvalidRuntime(String)

The runtime or runtime version specified is not supported.

InvalidSecurityGroupID(String)

The Security Group ID provided in the Lambda function VPC configuration is invalid.

InvalidSubnetID(String)

The Subnet ID provided in the Lambda function VPC configuration is invalid.

InvalidZipFile(String)

AWS Lambda could not unzip the function zip file.

KMSAccessDenied(String)

Lambda was unable to decrypt the environment variables because KMS access was denied. Check the Lambda function's KMS permissions.

KMSDisabled(String)

Lambda was unable to decrypt the environment variables because the KMS key used is disabled. Check the Lambda function's KMS key settings.

KMSInvalidState(String)

Lambda was unable to decrypt the environment variables because the KMS key used is in an invalid state for Decrypt. Check the function's KMS key settings.

KMSNotFound(String)

Lambda was unable to decrypt the environment variables because the KMS key was not found. Check the function's KMS key settings.

RequestTooLarge(String)

The request payload exceeded the Invoke request body JSON input limit. For more information, see Limits.

ResourceNotFound(String)

The resource (for example, a Lambda function or access policy statement) specified in the request does not exist.

Service(String)

The AWS Lambda service encountered an internal error.

SubnetIPAddressLimitReached(String)

AWS Lambda was not able to set up VPC access for the Lambda function because one or more configured subnets has no available IP addresses.

TooManyRequests(String)

UnsupportedMediaType(String)

The content type of the Invoke request body is not JSON.

HttpDispatch(HttpDispatchError)

An error occurred dispatching the HTTP request

Credentials(CredentialsError)

An error was encountered with AWS credentials.

Validation(String)

A validation error occurred. Details from AWS are provided.

Unknown(String)

An unknown error occurred. The raw HTTP response is provided.

Methods

impl InvokeError

fn from_body(body: &str) -> InvokeError