Struct rusoto_lambda::FunctionConfiguration

pub struct FunctionConfiguration {
    pub code_sha_256: Option<String>,
    pub code_size: Option<i64>,
    pub dead_letter_config: Option<DeadLetterConfig>,
    pub description: Option<String>,
    pub environment: Option<EnvironmentResponse>,
    pub function_arn: Option<String>,
    pub function_name: Option<String>,
    pub handler: Option<String>,
    pub kms_key_arn: Option<String>,
    pub last_modified: Option<String>,
    pub memory_size: Option<i64>,
    pub role: Option<String>,
    pub runtime: Option<String>,
    pub timeout: Option<i64>,
    pub tracing_config: Option<TracingConfigResponse>,
    pub version: Option<String>,
    pub vpc_config: Option<VpcConfigResponse>,
}

A complex type that describes function metadata.

Fields
code_sha_256: Option<String>

It is the SHA256 hash of your function deployment package.

code_size: Option<i64>

The size, in bytes, of the function .zip file you uploaded.

dead_letter_config: Option<DeadLetterConfig>

The parent object that contains the target ARN (Amazon Resource Name) of an Amazon SQS queue or Amazon SNS topic.

description: Option<String>

The user-provided description.

environment: Option<EnvironmentResponse>

The parent object that contains your environment's configuration settings.

function_arn: Option<String>

The Amazon Resource Name (ARN) assigned to the function.

function_name: Option<String>

The name of the function. Note that the length constraint applies only to the ARN. If you specify only the function name, it is limited to 64 characters in length.

handler: Option<String>

The function Lambda calls to begin executing your function.

kms_key_arn: Option<String>

The Amazon Resource Name (ARN) of the KMS key used to encrypt your function's environment variables. If empty, it means you are using the AWS Lambda default service key.

last_modified: Option<String>

The time stamp of the last time you updated the function. The time stamp is conveyed as a string complying with ISO-8601 in this way YYYY-MM-DDThh:mm:ssTZD (e.g., 1997-07-16T19:20:30+01:00). For more information, see Date and Time Formats.

memory_size: Option<i64>

The memory size, in MB, you configured for the function. Must be a multiple of 64 MB.

role: Option<String>

The Amazon Resource Name (ARN) of the IAM role that Lambda assumes when it executes your function to access any other Amazon Web Services (AWS) resources.

runtime: Option<String>

The runtime environment for the Lambda function.

timeout: Option<i64>

The function execution time at which Lambda should terminate the function. Because the execution time has cost implications, we recommend you set this value based on your expected execution time. The default is 3 seconds.

tracing_config: Option<TracingConfigResponse>

The parent object that contains your function's tracing settings.

version: Option<String>

The version of the Lambda function.

vpc_config: Option<VpcConfigResponse>

VPC configuration associated with your Lambda function.