Struct rusoto_lambda::LambdaClient

pub struct LambdaClient { /* fields omitted */ }

A client for the AWS Lambda API.

Methods

impl LambdaClient

pub fn new(region: Region) -> LambdaClient

Creates a client backed by the default tokio event loop.

The client will use the default credentials provider and tls client.

pub fn new_with<P, D>(
request_dispatcher: D,
credentials_provider: P,
region: Region
) -> LambdaClient where
P: ProvideAwsCredentials + Send + Sync + 'static,
P::Future: Send,
D: DispatchSignedRequest + Send + Sync + 'static,
D::Future: Send,