Struct rusoto_lambda::UpdateFunctionConfigurationRequest

pub struct UpdateFunctionConfigurationRequest {
    pub dead_letter_config: Option<DeadLetterConfig>,
    pub description: Option<String>,
    pub environment: Option<Environment>,
    pub function_name: String,
    pub handler: Option<String>,
    pub kms_key_arn: Option<String>,
    pub memory_size: Option<i64>,
    pub role: Option<String>,
    pub runtime: Option<String>,
    pub timeout: Option<i64>,
    pub tracing_config: Option<TracingConfig>,
    pub vpc_config: Option<VpcConfig>,
}

Fields
dead_letter_config: Option<DeadLetterConfig>

The parent object that contains the target ARN (Amazon Resource Name) of an Amazon SQS queue or Amazon SNS topic.

description: Option<String>

A short user-defined function description. AWS Lambda does not use this value. Assign a meaningful description as you see fit.

environment: Option<Environment>

The parent object that contains your environment's configuration settings.

function_name: String

The name of the Lambda function.

You can specify a function name (for example, Thumbnail) or you can specify Amazon Resource Name (ARN) of the function (for example, arn:aws:lambda:us-west-2:account-id:function:ThumbNail). AWS Lambda also allows you to specify a partial ARN (for example, account-id:Thumbnail). Note that the length constraint applies only to the ARN. If you specify only the function name, it is limited to 64 character in length.

handler: Option<String>

The function that Lambda calls to begin executing your function. For Node.js, it is the module-name.export value in your function.

kms_key_arn: Option<String>

The Amazon Resource Name (ARN) of the KMS key used to encrypt your function's environment variables. If you elect to use the AWS Lambda default service key, pass in an empty string ("") for this parameter.

memory_size: Option<i64>

The amount of memory, in MB, your Lambda function is given. AWS Lambda uses this memory size to infer the amount of CPU allocated to your function. Your function use-case determines your CPU and memory requirements. For example, a database operation might need less memory compared to an image processing function. The default value is 128 MB. The value must be a multiple of 64 MB.

role: Option<String>

The Amazon Resource Name (ARN) of the IAM role that Lambda will assume when it executes your function.

runtime: Option<String>

The runtime environment for the Lambda function.

To use the Python runtime v3.6, set the value to "python3.6". To use the Python runtime v2.7, set the value to "python2.7". To use the Node.js runtime v6.10, set the value to "nodejs6.10". To use the Node.js runtime v4.3, set the value to "nodejs4.3". To use the Python runtime v3.6, set the value to "python3.6". To use the Python runtime v2.7, set the value to "python2.7".

Node v0.10.42 is currently marked as deprecated. You must migrate existing functions to the newer Node.js runtime versions available on AWS Lambda (nodejs4.3 or nodejs6.10) as soon as possible. You can request a one-time extension until June 30, 2017 by going to the Lambda console and following the instructions provided. Failure to do so will result in an invalid parameter value error being returned. Note that you will have to follow this procedure for each region that contains functions written in the Node v0.10.42 runtime.

timeout: Option<i64>

The function execution time at which AWS Lambda should terminate the function. Because the execution time has cost implications, we recommend you set this value based on your expected execution time. The default is 3 seconds.

tracing_config: Option<TracingConfig>

The parent object that contains your function's tracing settings.

vpc_config: Option<VpcConfig>