Trait rusoto_lambda::Lambda

pub trait Lambda {
    fn add_permission(
        &self, 
        input: &AddPermissionRequest
    ) -> Result<AddPermissionResponse, AddPermissionError>;
    fn create_alias(
        &self, 
        input: &CreateAliasRequest
    ) -> Result<AliasConfiguration, CreateAliasError>;
    fn create_event_source_mapping(
        &self, 
        input: &CreateEventSourceMappingRequest
    ) -> Result<EventSourceMappingConfiguration, CreateEventSourceMappingError>;
    fn create_function(
        &self, 
        input: &CreateFunctionRequest
    ) -> Result<FunctionConfiguration, CreateFunctionError>;
    fn delete_alias(
        &self, 
        input: &DeleteAliasRequest
    ) -> Result<(), DeleteAliasError>;
    fn delete_event_source_mapping(
        &self, 
        input: &DeleteEventSourceMappingRequest
    ) -> Result<EventSourceMappingConfiguration, DeleteEventSourceMappingError>;
    fn delete_function(
        &self, 
        input: &DeleteFunctionRequest
    ) -> Result<(), DeleteFunctionError>;
    fn get_account_settings(
        &self
    ) -> Result<GetAccountSettingsResponse, GetAccountSettingsError>;
    fn get_alias(
        &self, 
        input: &GetAliasRequest
    ) -> Result<AliasConfiguration, GetAliasError>;
    fn get_event_source_mapping(
        &self, 
        input: &GetEventSourceMappingRequest
    ) -> Result<EventSourceMappingConfiguration, GetEventSourceMappingError>;
    fn get_function(
        &self, 
        input: &GetFunctionRequest
    ) -> Result<GetFunctionResponse, GetFunctionError>;
    fn get_function_configuration(
        &self, 
        input: &GetFunctionConfigurationRequest
    ) -> Result<FunctionConfiguration, GetFunctionConfigurationError>;
    fn get_policy(
        &self, 
        input: &GetPolicyRequest
    ) -> Result<GetPolicyResponse, GetPolicyError>;
    fn invoke(
        &self, 
        input: &InvocationRequest
    ) -> Result<InvocationResponse, InvokeError>;
    fn invoke_async(
        &self, 
        input: &InvokeAsyncRequest
    ) -> Result<InvokeAsyncResponse, InvokeAsyncError>;
    fn list_aliases(
        &self, 
        input: &ListAliasesRequest
    ) -> Result<ListAliasesResponse, ListAliasesError>;
    fn list_event_source_mappings(
        &self, 
        input: &ListEventSourceMappingsRequest
    ) -> Result<ListEventSourceMappingsResponse, ListEventSourceMappingsError>;
    fn list_functions(
        &self, 
        input: &ListFunctionsRequest
    ) -> Result<ListFunctionsResponse, ListFunctionsError>;
    fn list_tags(
        &self, 
        input: &ListTagsRequest
    ) -> Result<ListTagsResponse, ListTagsError>;
    fn list_versions_by_function(
        &self, 
        input: &ListVersionsByFunctionRequest
    ) -> Result<ListVersionsByFunctionResponse, ListVersionsByFunctionError>;
    fn publish_version(
        &self, 
        input: &PublishVersionRequest
    ) -> Result<FunctionConfiguration, PublishVersionError>;
    fn remove_permission(
        &self, 
        input: &RemovePermissionRequest
    ) -> Result<(), RemovePermissionError>;
    fn tag_resource(
        &self, 
        input: &TagResourceRequest
    ) -> Result<(), TagResourceError>;
    fn untag_resource(
        &self, 
        input: &UntagResourceRequest
    ) -> Result<(), UntagResourceError>;
    fn update_alias(
        &self, 
        input: &UpdateAliasRequest
    ) -> Result<AliasConfiguration, UpdateAliasError>;
    fn update_event_source_mapping(
        &self, 
        input: &UpdateEventSourceMappingRequest
    ) -> Result<EventSourceMappingConfiguration, UpdateEventSourceMappingError>;
    fn update_function_code(
        &self, 
        input: &UpdateFunctionCodeRequest
    ) -> Result<FunctionConfiguration, UpdateFunctionCodeError>;
    fn update_function_configuration(
        &self, 
        input: &UpdateFunctionConfigurationRequest
    ) -> Result<FunctionConfiguration, UpdateFunctionConfigurationError>;
}

Trait representing the capabilities of the AWS Lambda API. AWS Lambda clients implement this trait.

Required Methods

fn add_permission(
&self,
input: &AddPermissionRequest
) -> Result<AddPermissionResponse, AddPermissionError>

Adds a permission to the resource policy associated with the specified AWS Lambda function. You use resource policies to grant permissions to event sources that use push model. In a push model, event sources (such as Amazon S3 and custom applications) invoke your Lambda function. Each permission you add to the resource policy allows an event source, permission to invoke the Lambda function.

For information about the push model, see AWS Lambda: How it Works.

If you are using versioning, the permissions you add are specific to the Lambda function version or alias you specify in the AddPermission request via the Qualifier parameter. For more information about versioning, see AWS Lambda Function Versioning and Aliases.

This operation requires permission for the lambda:AddPermission action.

fn create_alias(
&self,
input: &CreateAliasRequest
) -> Result<AliasConfiguration, CreateAliasError>

Creates an alias that points to the specified Lambda function version. For more information, see Introduction to AWS Lambda Aliases.

Alias names are unique for a given function. This requires permission for the lambda:CreateAlias action.

fn create_event_source_mapping(
&self,
input: &CreateEventSourceMappingRequest
) -> Result<EventSourceMappingConfiguration, CreateEventSourceMappingError>

Identifies a stream as an event source for a Lambda function. It can be either an Amazon Kinesis stream or an Amazon DynamoDB stream. AWS Lambda invokes the specified function when records are posted to the stream.

This association between a stream source and a Lambda function is called the event source mapping.

This event source mapping is relevant only in the AWS Lambda pull model, where AWS Lambda invokes the function. For more information, see AWS Lambda: How it Works in the AWS Lambda Developer Guide.

You provide mapping information (for example, which stream to read from and which Lambda function to invoke) in the request body.

Each event source, such as an Amazon Kinesis or a DynamoDB stream, can be associated with multiple AWS Lambda function. A given Lambda function can be associated with multiple AWS event sources.

If you are using versioning, you can specify a specific function version or an alias via the function name parameter. For more information about versioning, see AWS Lambda Function Versioning and Aliases.

This operation requires permission for the lambda:CreateEventSourceMapping action.

fn create_function(
&self,
input: &CreateFunctionRequest
) -> Result<FunctionConfiguration, CreateFunctionError>

Creates a new Lambda function. The function metadata is created from the request parameters, and the code for the function is provided by a .zip file in the request body. If the function name already exists, the operation will fail. Note that the function name is case-sensitive.

If you are using versioning, you can also publish a version of the Lambda function you are creating using the Publish parameter. For more information about versioning, see AWS Lambda Function Versioning and Aliases.

This operation requires permission for the lambda:CreateFunction action.

fn delete_alias(
&self,
input: &DeleteAliasRequest
) -> Result<(), DeleteAliasError>

Deletes the specified Lambda function alias. For more information, see Introduction to AWS Lambda Aliases.

This requires permission for the lambda:DeleteAlias action.

fn delete_event_source_mapping(
&self,
input: &DeleteEventSourceMappingRequest
) -> Result<EventSourceMappingConfiguration, DeleteEventSourceMappingError>

Removes an event source mapping. This means AWS Lambda will no longer invoke the function for events in the associated source.

This operation requires permission for the lambda:DeleteEventSourceMapping action.

fn delete_function(
&self,
input: &DeleteFunctionRequest
) -> Result<(), DeleteFunctionError>

Deletes the specified Lambda function code and configuration.

If you are using the versioning feature and you don't specify a function version in your DeleteFunction request, AWS Lambda will delete the function, including all its versions, and any aliases pointing to the function versions. To delete a specific function version, you must provide the function version via the Qualifier parameter. For information about function versioning, see AWS Lambda Function Versioning and Aliases.

When you delete a function the associated resource policy is also deleted. You will need to delete the event source mappings explicitly.

This operation requires permission for the lambda:DeleteFunction action.

fn get_account_settings(
&self
) -> Result<GetAccountSettingsResponse, GetAccountSettingsError>

Returns a customer's account settings.

You can use this operation to retrieve Lambda limits information, such as code size and concurrency limits. For more information about limits, see AWS Lambda Limits. You can also retrieve resource usage statistics, such as code storage usage and function count.

fn get_alias(
&self,
input: &GetAliasRequest
) -> Result<AliasConfiguration, GetAliasError>

Returns the specified alias information such as the alias ARN, description, and function version it is pointing to. For more information, see Introduction to AWS Lambda Aliases.

This requires permission for the lambda:GetAlias action.

fn get_event_source_mapping(
&self,
input: &GetEventSourceMappingRequest
) -> Result<EventSourceMappingConfiguration, GetEventSourceMappingError>

Returns configuration information for the specified event source mapping (see CreateEventSourceMapping).

This operation requires permission for the lambda:GetEventSourceMapping action.

fn get_function(
&self,
input: &GetFunctionRequest
) -> Result<GetFunctionResponse, GetFunctionError>

Returns the configuration information of the Lambda function and a presigned URL link to the .zip file you uploaded with CreateFunction so you can download the .zip file. Note that the URL is valid for up to 10 minutes. The configuration information is the same information you provided as parameters when uploading the function.

Using the optional Qualifier parameter, you can specify a specific function version for which you want this information. If you don't specify this parameter, the API uses unqualified function ARN which return information about the $LATEST version of the Lambda function. For more information, see AWS Lambda Function Versioning and Aliases.

This operation requires permission for the lambda:GetFunction action.

fn get_function_configuration(
&self,
input: &GetFunctionConfigurationRequest
) -> Result<FunctionConfiguration, GetFunctionConfigurationError>

Returns the configuration information of the Lambda function. This the same information you provided as parameters when uploading the function by using CreateFunction.

If you are using the versioning feature, you can retrieve this information for a specific function version by using the optional Qualifier parameter and specifying the function version or alias that points to it. If you don't provide it, the API returns information about the $LATEST version of the function. For more information about versioning, see AWS Lambda Function Versioning and Aliases.

This operation requires permission for the lambda:GetFunctionConfiguration operation.

fn get_policy(
&self,
input: &GetPolicyRequest
) -> Result<GetPolicyResponse, GetPolicyError>

Returns the resource policy associated with the specified Lambda function.

If you are using the versioning feature, you can get the resource policy associated with the specific Lambda function version or alias by specifying the version or alias name using the Qualifier parameter. For more information about versioning, see AWS Lambda Function Versioning and Aliases.

You need permission for the lambda:GetPolicy action.

fn invoke(
&self,
input: &InvocationRequest
) -> Result<InvocationResponse, InvokeError>

Invokes a specific Lambda function. For an example, see Create the Lambda Function and Test It Manually.

If you are using the versioning feature, you can invoke the specific function version by providing function version or alias name that is pointing to the function version using the Qualifier parameter in the request. If you don't provide the Qualifier parameter, the $LATEST version of the Lambda function is invoked. Invocations occur at least once in response to an event and functions must be idempotent to handle this. For information about the versioning feature, see AWS Lambda Function Versioning and Aliases.

This operation requires permission for the lambda:InvokeFunction action.

fn invoke_async(
&self,
input: &InvokeAsyncRequest
) -> Result<InvokeAsyncResponse, InvokeAsyncError>

This API is deprecated. We recommend you use Invoke API (see Invoke).

Submits an invocation request to AWS Lambda. Upon receiving the request, Lambda executes the specified function asynchronously. To see the logs generated by the Lambda function execution, see the CloudWatch Logs console.

This operation requires permission for the lambda:InvokeFunction action.

fn list_aliases(
&self,
input: &ListAliasesRequest
) -> Result<ListAliasesResponse, ListAliasesError>

Returns list of aliases created for a Lambda function. For each alias, the response includes information such as the alias ARN, description, alias name, and the function version to which it points. For more information, see Introduction to AWS Lambda Aliases.

This requires permission for the lambda:ListAliases action.

fn list_event_source_mappings(
&self,
input: &ListEventSourceMappingsRequest
) -> Result<ListEventSourceMappingsResponse, ListEventSourceMappingsError>

Returns a list of event source mappings you created using the CreateEventSourceMapping (see CreateEventSourceMapping).

For each mapping, the API returns configuration information. You can optionally specify filters to retrieve specific event source mappings.

If you are using the versioning feature, you can get list of event source mappings for a specific Lambda function version or an alias as described in the FunctionName parameter. For information about the versioning feature, see AWS Lambda Function Versioning and Aliases.

This operation requires permission for the lambda:ListEventSourceMappings action.

fn list_functions(
&self,
input: &ListFunctionsRequest
) -> Result<ListFunctionsResponse, ListFunctionsError>

Returns a list of your Lambda functions. For each function, the response includes the function configuration information. You must use GetFunction to retrieve the code for your function.

This operation requires permission for the lambda:ListFunctions action.

If you are using versioning feature, the response returns list of $LATEST versions of your functions. For information about the versioning feature, see AWS Lambda Function Versioning and Aliases.

fn list_tags(
&self,
input: &ListTagsRequest
) -> Result<ListTagsResponse, ListTagsError>

Returns a list of tags assigned to a function when supplied the function ARN (Amazon Resource Name).

fn list_versions_by_function(
&self,
input: &ListVersionsByFunctionRequest
) -> Result<ListVersionsByFunctionResponse, ListVersionsByFunctionError>

List all versions of a function. For information about the versioning feature, see AWS Lambda Function Versioning and Aliases.

fn publish_version(
&self,
input: &PublishVersionRequest
) -> Result<FunctionConfiguration, PublishVersionError>

Publishes a version of your function from the current snapshot of $LATEST. That is, AWS Lambda takes a snapshot of the function code and configuration information from $LATEST and publishes a new version. The code and configuration cannot be modified after publication. For information about the versioning feature, see AWS Lambda Function Versioning and Aliases.

fn remove_permission(
&self,
input: &RemovePermissionRequest
) -> Result<(), RemovePermissionError>

You can remove individual permissions from an resource policy associated with a Lambda function by providing a statement ID that you provided when you added the permission.

If you are using versioning, the permissions you remove are specific to the Lambda function version or alias you specify in the AddPermission request via the Qualifier parameter. For more information about versioning, see AWS Lambda Function Versioning and Aliases.

Note that removal of a permission will cause an active event source to lose permission to the function.

You need permission for the lambda:RemovePermission action.

fn tag_resource(
&self,
input: &TagResourceRequest
) -> Result<(), TagResourceError>

Creates a list of tags (key-value pairs) on the Lambda function. Requires the Lambda function ARN (Amazon Resource Name). If a key is specified without a value, Lambda creates a tag with the specified key and a value of null.

fn untag_resource(
&self,
input: &UntagResourceRequest
) -> Result<(), UntagResourceError>

Removes tags from a Lambda function. Requires the function ARN (Amazon Resource Name).

fn update_alias(
&self,
input: &UpdateAliasRequest
) -> Result<AliasConfiguration, UpdateAliasError>

Using this API you can update the function version to which the alias points and the alias description. For more information, see Introduction to AWS Lambda Aliases.

This requires permission for the lambda:UpdateAlias action.

fn update_event_source_mapping(
&self,
input: &UpdateEventSourceMappingRequest
) -> Result<EventSourceMappingConfiguration, UpdateEventSourceMappingError>

You can update an event source mapping. This is useful if you want to change the parameters of the existing mapping without losing your position in the stream. You can change which function will receive the stream records, but to change the stream itself, you must create a new mapping.

If you are using the versioning feature, you can update the event source mapping to map to a specific Lambda function version or alias as described in the FunctionName parameter. For information about the versioning feature, see AWS Lambda Function Versioning and Aliases.

If you disable the event source mapping, AWS Lambda stops polling. If you enable again, it will resume polling from the time it had stopped polling, so you don't lose processing of any records. However, if you delete event source mapping and create it again, it will reset.

This operation requires permission for the lambda:UpdateEventSourceMapping action.

fn update_function_code(
&self,
input: &UpdateFunctionCodeRequest
) -> Result<FunctionConfiguration, UpdateFunctionCodeError>

Updates the code for the specified Lambda function. This operation must only be used on an existing Lambda function and cannot be used to update the function configuration.

If you are using the versioning feature, note this API will always update the $LATEST version of your Lambda function. For information about the versioning feature, see AWS Lambda Function Versioning and Aliases.

This operation requires permission for the lambda:UpdateFunctionCode action.

fn update_function_configuration(
&self,
input: &UpdateFunctionConfigurationRequest
) -> Result<FunctionConfiguration, UpdateFunctionConfigurationError>

Updates the configuration parameters for the specified Lambda function by using the values provided in the request. You provide only the parameters you want to change. This operation must only be used on an existing Lambda function and cannot be used to update the function's code.

If you are using the versioning feature, note this API will always update the $LATEST version of your Lambda function. For information about the versioning feature, see AWS Lambda Function Versioning and Aliases.

This operation requires permission for the lambda:UpdateFunctionConfiguration action.