Crate rusoto_lambda

AWS Lambda
Overview

This is the AWS Lambda API Reference. The AWS Lambda Developer Guide provides additional information. For the service overview, see What is AWS Lambda, and for information about how the service works, see AWS Lambda: How it Works in the AWS Lambda Developer Guide.

If you're using the service, you're probably looking for LambdaClient and Lambda.

Structs

AccountLimit
Provides limits of code size and concurrency associated with the current account and region.

AccountUsage
Provides code size usage and function count associated with the current account and region.

AddPermissionRequest

AddPermissionResponse

AliasConfiguration
Provides configuration information about a Lambda function version alias.

AliasRoutingConfiguration
The parent object that implements what percentage of traffic will invoke each function version. For more information, see lambda-traffic-shifting-using-aliases.

Concurrency

CreateAliasRequest

CreateEventSourceMappingRequest

CreateFunctionRequest

DeadLetterConfig
The Amazon Resource Name (ARN) of an Amazon SQS queue or Amazon SNS topic you specify as your Dead Letter Queue (DLQ). For more information, see dlq.

DeleteAliasRequest

DeleteEventSourceMappingRequest

DeleteFunctionConcurrencyRequest

DeleteFunctionRequest

Environment
The parent object that contains your environment's configuration settings.

EnvironmentError
The parent object that contains error information associated with your configuration settings.

EnvironmentResponse
The parent object returned that contains your environment's configuration settings or any error information associated with your configuration settings.

EventSourceMappingConfiguration
Describes mapping between an Amazon Kinesis or DynamoDB stream or an Amazon SQS queue and a Lambda function.

FunctionCode
The code for the Lambda function.

FunctionCodeLocation
The object for the Lambda function location.

FunctionConfiguration
A complex type that describes function metadata.

GetAccountSettingsRequest

GetAccountSettingsResponse

GetAliasRequest

GetEventSourceMappingRequest

GetFunctionConfigurationRequest

GetFunctionRequest

GetFunctionResponse
This response contains the object for the Lambda function location (see FunctionCodeLocation.

GetPolicyRequest

GetPolicyResponse

InvocationRequest

InvocationResponse
Upon success, returns an empty response. Otherwise, throws an exception.

InvokeAsyncRequest

InvokeAsyncResponse
Upon success, it returns empty response. Otherwise, throws an exception.

LambdaClient
A client for the AWS Lambda API.

ListAliasesRequest

ListAliasesResponse

ListEventSourceMappingsRequest

ListEventSourceMappingsResponse
Contains a list of event sources (see EventSourceMappingConfiguration)

ListFunctionsRequest

ListFunctionsResponse
Contains a list of AWS Lambda function configurations (see FunctionConfiguration.

ListTagsRequest

ListTagsResponse

ListVersionsByFunctionRequest

ListVersionsByFunctionResponse

PublishVersionRequest

PutFunctionConcurrencyRequest

RemovePermissionRequest

TagResourceRequest

TracingConfig
The parent object that contains your function's tracing settings.

TracingConfigResponse
Parent object of the tracing information associated with your Lambda function.

UntagResourceRequest

UpdateAliasRequest

UpdateEventSourceMappingRequest

UpdateFunctionCodeRequest

UpdateFunctionConfigurationRequest

VpcConfig
If your Lambda function accesses resources in a VPC, you provide this parameter identifying the list of security group IDs and subnet IDs. These must belong to the same VPC. You must provide at least one security group and one subnet ID.

VpcConfigResponse
VPC configuration associated with your Lambda function.

Enums

AddPermissionError
Errors returned by AddPermission

CreateAliasError
Errors returned by CreateAlias

CreateEventSourceMappingError
Errors returned by CreateEventSourceMapping

CreateFunctionError
Errors returned by CreateFunction

DeleteAliasError
Errors returned by DeleteAlias

DeleteEventSourceMappingError
Errors returned by DeleteEventSourceMapping

DeleteFunctionConcurrencyError
Errors returned by DeleteFunctionConcurrency

DeleteFunctionError
Errors returned by DeleteFunction

GetAccountSettingsError
Errors returned by GetAccountSettings

GetAliasError
Errors returned by GetAlias

GetEventSourceMappingError
Errors returned by GetEventSourceMapping

GetFunctionConfigurationError
Errors returned by GetFunctionConfiguration

GetFunctionError
Errors returned by GetFunction

GetPolicyError
Errors returned by GetPolicy

InvokeAsyncError
Errors returned by InvokeAsync

InvokeError
Errors returned by Invoke

ListAliasesError
Errors returned by ListAliases

ListEventSourceMappingsError
Errors returned by ListEventSourceMappings

ListFunctionsError
Errors returned by ListFunctions

ListTagsError
Errors returned by ListTags

ListVersionsByFunctionError
Errors returned by ListVersionsByFunction

PublishVersionError
Errors returned by PublishVersion

PutFunctionConcurrencyError
Errors returned by PutFunctionConcurrency

RemovePermissionError
Errors returned by RemovePermission

TagResourceError
Errors returned by TagResource

UntagResourceError
Errors returned by UntagResource

UpdateAliasError
Errors returned by UpdateAlias

UpdateEventSourceMappingError
Errors returned by UpdateEventSourceMapping

UpdateFunctionCodeError
Errors returned by UpdateFunctionCode

UpdateFunctionConfigurationError
Errors returned by UpdateFunctionConfiguration

Traits

Lambda
Trait representing the capabilities of the AWS Lambda API. AWS Lambda clients implement this trait.