Struct rusoto_lambda::CreateFunctionRequest

pub struct CreateFunctionRequest {
    pub code: FunctionCode,
    pub dead_letter_config: Option<DeadLetterConfig>,
    pub description: Option<Description>,
    pub environment: Option<Environment>,
    pub function_name: FunctionName,
    pub handler: Handler,
    pub kms_key_arn: Option<KMSKeyArn>,
    pub memory_size: Option<MemorySize>,
    pub publish: Option<Boolean>,
    pub role: RoleArn,
    pub runtime: Runtime,
    pub tags: Option<Tags>,
    pub timeout: Option<Timeout>,
    pub tracing_config: Option<TracingConfig>,
    pub vpc_config: Option<VpcConfig>,
}

Fields
code: FunctionCode

The code for the Lambda function.

dead_letter_config: Option<DeadLetterConfig>

The parent object that contains the target ARN (Amazon Resource Name) of an Amazon SQS queue or Amazon SNS topic.

description: Option<Description>

A short, user-defined function description. Lambda does not use this value. Assign a meaningful description as you see fit.

environment: Option<Environment> function_name: FunctionName

The name you want to assign to the function you are uploading. The function names appear in the console and are returned in the ListFunctions API. Function names are used to specify functions to other AWS Lambda API operations, such as Invoke. Note that the length constraint applies only to the ARN. If you specify only the function name, it is limited to 64 characters in length.

handler: Handler

The function within your code that Lambda calls to begin execution. For Node.js, it is the module-name.export value in your function. For Java, it can be package.class-name::handler or package.class-name. For more information, see Lambda Function Handler (Java).

kms_key_arn: Option<KMSKeyArn>

The Amazon Resource Name (ARN) of the KMS key used to encrypt your function's environment variables. If not provided, AWS Lambda will use a default service key.

memory_size: Option<MemorySize>

The amount of memory, in MB, your Lambda function is given. Lambda uses this memory size to infer the amount of CPU and memory allocated to your function. Your function use-case determines your CPU and memory requirements. For example, a database operation might need less memory compared to an image processing function. The default value is 128 MB. The value must be a multiple of 64 MB.

publish: Option<Boolean>

This boolean parameter can be used to request AWS Lambda to create the Lambda function and publish a version as an atomic operation.

role: RoleArn

The Amazon Resource Name (ARN) of the IAM role that Lambda assumes when it executes your function to access any other Amazon Web Services (AWS) resources. For more information, see AWS Lambda: How it Works.

runtime: Runtime

The runtime environment for the Lambda function you are uploading.

To use the Python runtime v3.6, set the value to "python3.6". To use the Python runtime v2.7, set the value to "python2.7". To use the Node.js runtime v6.10, set the value to "nodejs6.10". To use the Node.js runtime v4.3, set the value to "nodejs4.3".

You can no longer create functions using the v0.10.42 runtime version as of November, 2016. Existing functions will be supported until early 2017, but we recommend you migrate them to either nodejs6.10 or nodejs4.3 runtime version as soon as possible.

tags: Option<Tags>

The list of tags (key-value pairs) assigned to the new function.

timeout: Option<Timeout>

The function execution time at which Lambda should terminate the function. Because the execution time has cost implications, we recommend you set this value based on your expected execution time. The default is 3 seconds.

tracing_config: Option<TracingConfig>

The parent object that contains your function's tracing settings.

vpc_config: Option<VpcConfig>

If your Lambda function accesses resources in a VPC, you provide this parameter identifying the list of security group IDs and subnet IDs. These must belong to the same VPC. You must provide at least one security group and one subnet ID.