Trait rusoto_lambda::Lambda

pub trait Lambda {
    fn add_layer_version_permission(
        &self, 
        input: AddLayerVersionPermissionRequest
    ) -> RusotoFuture<AddLayerVersionPermissionResponse, AddLayerVersionPermissionError>;

    fn add_permission(
        &self, 
        input: AddPermissionRequest
    ) -> RusotoFuture<AddPermissionResponse, AddPermissionError>;

    fn create_alias(
        &self, 
        input: CreateAliasRequest
    ) -> RusotoFuture<AliasConfiguration, CreateAliasError>;

    fn create_event_source_mapping(
        &self, 
        input: CreateEventSourceMappingRequest
    ) -> RusotoFuture<EventSourceMappingConfiguration, CreateEventSourceMappingError>;

    fn create_function(
        &self, 
        input: CreateFunctionRequest
    ) -> RusotoFuture<FunctionConfiguration, CreateFunctionError>;

    fn delete_alias(
        &self, 
        input: DeleteAliasRequest
    ) -> RusotoFuture<(), DeleteAliasError>;

    fn delete_event_source_mapping(
        &self, 
        input: DeleteEventSourceMappingRequest
    ) -> RusotoFuture<EventSourceMappingConfiguration, DeleteEventSourceMappingError>;

    fn delete_function(
        &self, 
        input: DeleteFunctionRequest
    ) -> RusotoFuture<(), DeleteFunctionError>;

    fn delete_function_concurrency(
        &self, 
        input: DeleteFunctionConcurrencyRequest
    ) -> RusotoFuture<(), DeleteFunctionConcurrencyError>;

    fn delete_layer_version(
        &self, 
        input: DeleteLayerVersionRequest
    ) -> RusotoFuture<(), DeleteLayerVersionError>;

    fn get_account_settings(
        &self
    ) -> RusotoFuture<GetAccountSettingsResponse, GetAccountSettingsError>;

    fn get_alias(
        &self, 
        input: GetAliasRequest
    ) -> RusotoFuture<AliasConfiguration, GetAliasError>;

    fn get_event_source_mapping(
        &self, 
        input: GetEventSourceMappingRequest
    ) -> RusotoFuture<EventSourceMappingConfiguration, GetEventSourceMappingError>;

    fn get_function(
        &self, 
        input: GetFunctionRequest
    ) -> RusotoFuture<GetFunctionResponse, GetFunctionError>;

    fn get_function_configuration(
        &self, 
        input: GetFunctionConfigurationRequest
    ) -> RusotoFuture<FunctionConfiguration, GetFunctionConfigurationError>;

    fn get_layer_version(
        &self, 
        input: GetLayerVersionRequest
    ) -> RusotoFuture<GetLayerVersionResponse, GetLayerVersionError>;

    fn get_layer_version_by_arn(
        &self, 
        input: GetLayerVersionByArnRequest
    ) -> RusotoFuture<GetLayerVersionResponse, GetLayerVersionByArnError>;

    fn get_layer_version_policy(
        &self, 
        input: GetLayerVersionPolicyRequest
    ) -> RusotoFuture<GetLayerVersionPolicyResponse, GetLayerVersionPolicyError>;

    fn get_policy(
        &self, 
        input: GetPolicyRequest
    ) -> RusotoFuture<GetPolicyResponse, GetPolicyError>;

    fn invoke(
        &self, 
        input: InvocationRequest
    ) -> RusotoFuture<InvocationResponse, InvokeError>;

    fn invoke_async(
        &self, 
        input: InvokeAsyncRequest
    ) -> RusotoFuture<InvokeAsyncResponse, InvokeAsyncError>;

    fn list_aliases(
        &self, 
        input: ListAliasesRequest
    ) -> RusotoFuture<ListAliasesResponse, ListAliasesError>;

    fn list_event_source_mappings(
        &self, 
        input: ListEventSourceMappingsRequest
    ) -> RusotoFuture<ListEventSourceMappingsResponse, ListEventSourceMappingsError>;

    fn list_functions(
        &self, 
        input: ListFunctionsRequest
    ) -> RusotoFuture<ListFunctionsResponse, ListFunctionsError>;

    fn list_layer_versions(
        &self, 
        input: ListLayerVersionsRequest
    ) -> RusotoFuture<ListLayerVersionsResponse, ListLayerVersionsError>;

    fn list_layers(
        &self, 
        input: ListLayersRequest
    ) -> RusotoFuture<ListLayersResponse, ListLayersError>;

    fn list_tags(
        &self, 
        input: ListTagsRequest
    ) -> RusotoFuture<ListTagsResponse, ListTagsError>;

    fn list_versions_by_function(
        &self, 
        input: ListVersionsByFunctionRequest
    ) -> RusotoFuture<ListVersionsByFunctionResponse, ListVersionsByFunctionError>;

    fn publish_layer_version(
        &self, 
        input: PublishLayerVersionRequest
    ) -> RusotoFuture<PublishLayerVersionResponse, PublishLayerVersionError>;

    fn publish_version(
        &self, 
        input: PublishVersionRequest
    ) -> RusotoFuture<FunctionConfiguration, PublishVersionError>;

    fn put_function_concurrency(
        &self, 
        input: PutFunctionConcurrencyRequest
    ) -> RusotoFuture<Concurrency, PutFunctionConcurrencyError>;

    fn remove_layer_version_permission(
        &self, 
        input: RemoveLayerVersionPermissionRequest
    ) -> RusotoFuture<(), RemoveLayerVersionPermissionError>;

    fn remove_permission(
        &self, 
        input: RemovePermissionRequest
    ) -> RusotoFuture<(), RemovePermissionError>;

    fn tag_resource(
        &self, 
        input: TagResourceRequest
    ) -> RusotoFuture<(), TagResourceError>;

    fn untag_resource(
        &self, 
        input: UntagResourceRequest
    ) -> RusotoFuture<(), UntagResourceError>;

    fn update_alias(
        &self, 
        input: UpdateAliasRequest
    ) -> RusotoFuture<AliasConfiguration, UpdateAliasError>;

    fn update_event_source_mapping(
        &self, 
        input: UpdateEventSourceMappingRequest
    ) -> RusotoFuture<EventSourceMappingConfiguration, UpdateEventSourceMappingError>;

    fn update_function_code(
        &self, 
        input: UpdateFunctionCodeRequest
    ) -> RusotoFuture<FunctionConfiguration, UpdateFunctionCodeError>;

    fn update_function_configuration(
        &self, 
        input: UpdateFunctionConfigurationRequest
    ) -> RusotoFuture<FunctionConfiguration, UpdateFunctionConfigurationError>;
}

Trait representing the capabilities of the AWS Lambda API. AWS Lambda clients implement this trait.

Required methods

fn add_layer_version_permission(
&self,
input: AddLayerVersionPermissionRequest
) -> RusotoFuture<AddLayerVersionPermissionResponse, AddLayerVersionPermissionError>

Adds permissions to the resource-based policy of a version of an AWS Lambda layer. Use this action to grant layer usage permission to other accounts. You can grant permission to a single account, all AWS accounts, or all accounts in an organization.

To revoke permission, call RemoveLayerVersionPermission with the statement ID that you specified when you added it.

fn add_permission(
&self,
input: AddPermissionRequest
) -> RusotoFuture<AddPermissionResponse, AddPermissionError>

Grants an AWS service or another account permission to use a function. You can apply the policy at the function level, or specify a qualifier to restrict access to a single version or alias. If you use a qualifier, the invoker must use the full Amazon Resource Name (ARN) of that version or alias to invoke the function.

To grant permission to another account, specify the account ID as the Principal. For AWS services, the principal is a domain-style identifier defined by the service, like s3.amazonaws.com or sns.amazonaws.com. For AWS services, you can also specify the ARN or owning account of the associated resource as the SourceArn or SourceAccount. If you grant permission to a service principal without specifying the source, other accounts could potentially configure resources in their account to invoke your Lambda function.

This action adds a statement to a resource-based permission policy for the function. For more information about function policies, see Lambda Function Policies.

fn create_alias(
&self,
input: CreateAliasRequest
) -> RusotoFuture<AliasConfiguration, CreateAliasError>

Creates an alias for a Lambda function version. Use aliases to provide clients with a function identifier that you can update to invoke a different version.

You can also map an alias to split invocation requests between two versions. Use the RoutingConfig parameter to specify a second version and the percentage of invocation requests that it receives.

fn create_event_source_mapping(
&self,
input: CreateEventSourceMappingRequest
) -> RusotoFuture<EventSourceMappingConfiguration, CreateEventSourceMappingError>

Creates a mapping between an event source and an AWS Lambda function. Lambda reads items from the event source and triggers the function.

For details about each event source type, see the following topics.

Using AWS Lambda with Amazon Kinesis

Using AWS Lambda with Amazon SQS

Using AWS Lambda with Amazon DynamoDB

fn create_function(
&self,
input: CreateFunctionRequest
) -> RusotoFuture<FunctionConfiguration, CreateFunctionError>

Creates a Lambda function. To create a function, you need a deployment package and an execution role. The deployment package contains your function code. The execution role grants the function permission to use AWS services, such as Amazon CloudWatch Logs for log streaming and AWS X-Ray for request tracing.

A function has an unpublished version, and can have published versions and aliases. The unpublished version changes when you update your function's code and configuration. A published version is a snapshot of your function code and configuration that can't be changed. An alias is a named resource that maps to a version, and can be changed to map to a different version. Use the Publish parameter to create version 1 of your function from its initial configuration.

The other parameters let you configure version-specific and function-level settings. You can modify version-specific settings later with UpdateFunctionConfiguration. Function-level settings apply to both the unpublished and published versions of the function, and include tags (TagResource) and per-function concurrency limits (PutFunctionConcurrency).

If another account or an AWS service invokes your function, use AddPermission to grant permission by creating a resource-based IAM policy. You can grant permissions at the function level, on a version, or on an alias.

To invoke your function directly, use Invoke. To invoke your function in response to events in other AWS services, create an event source mapping (CreateEventSourceMapping), or configure a function trigger in the other service. For more information, see Invoking Functions.

fn delete_alias(
&self,
input: DeleteAliasRequest
) -> RusotoFuture<(), DeleteAliasError>

Deletes a Lambda function alias.

fn delete_event_source_mapping(
&self,
input: DeleteEventSourceMappingRequest
) -> RusotoFuture<EventSourceMappingConfiguration, DeleteEventSourceMappingError>

Deletes an event source mapping. You can get the identifier of a mapping from the output of ListEventSourceMappings.

fn delete_function(
&self,
input: DeleteFunctionRequest
) -> RusotoFuture<(), DeleteFunctionError>

Deletes a Lambda function. To delete a specific function version, use the Qualifier parameter. Otherwise, all versions and aliases are deleted.

To delete Lambda event source mappings that invoke a function, use DeleteEventSourceMapping. For AWS services and resources that invoke your function directly, delete the trigger in the service where you originally configured it.

fn delete_function_concurrency(
&self,
input: DeleteFunctionConcurrencyRequest
) -> RusotoFuture<(), DeleteFunctionConcurrencyError>

Removes a concurrent execution limit from a function.

fn delete_layer_version(
&self,
input: DeleteLayerVersionRequest
) -> RusotoFuture<(), DeleteLayerVersionError>

Deletes a version of an AWS Lambda layer. Deleted versions can no longer be viewed or added to functions. To avoid breaking functions, a copy of the version remains in Lambda until no functions refer to it.

fn get_account_settings(
&self
) -> RusotoFuture<GetAccountSettingsResponse, GetAccountSettingsError>

Retrieves details about your account's limits and usage in an AWS Region.

fn get_alias(
&self,
input: GetAliasRequest
) -> RusotoFuture<AliasConfiguration, GetAliasError>

Returns details about a Lambda function alias.

fn get_event_source_mapping(
&self,
input: GetEventSourceMappingRequest
) -> RusotoFuture<EventSourceMappingConfiguration, GetEventSourceMappingError>

Returns details about an event source mapping. You can get the identifier of a mapping from the output of ListEventSourceMappings.

fn get_function(
&self,
input: GetFunctionRequest
) -> RusotoFuture<GetFunctionResponse, GetFunctionError>

Returns information about the function or function version, with a link to download the deployment package that's valid for 10 minutes. If you specify a function version, only details that are specific to that version are returned.

fn get_function_configuration(
&self,
input: GetFunctionConfigurationRequest
) -> RusotoFuture<FunctionConfiguration, GetFunctionConfigurationError>

Returns the version-specific settings of a Lambda function or version. The output includes only options that can vary between versions of a function. To modify these settings, use UpdateFunctionConfiguration.

To get all of a function's details, including function-level settings, use GetFunction.

fn get_layer_version(
&self,
input: GetLayerVersionRequest
) -> RusotoFuture<GetLayerVersionResponse, GetLayerVersionError>

Returns information about a version of an AWS Lambda layer, with a link to download the layer archive that's valid for 10 minutes.

fn get_layer_version_by_arn(
&self,
input: GetLayerVersionByArnRequest
) -> RusotoFuture<GetLayerVersionResponse, GetLayerVersionByArnError>

Returns information about a version of an AWS Lambda layer, with a link to download the layer archive that's valid for 10 minutes.

fn get_layer_version_policy(
&self,
input: GetLayerVersionPolicyRequest
) -> RusotoFuture<GetLayerVersionPolicyResponse, GetLayerVersionPolicyError>

Returns the permission policy for a version of an AWS Lambda layer. For more information, see AddLayerVersionPermission.

fn get_policy(
&self,
input: GetPolicyRequest
) -> RusotoFuture<GetPolicyResponse, GetPolicyError>

Returns the resource-based IAM policy for a function, version, or alias.

fn invoke(
&self,
input: InvocationRequest
) -> RusotoFuture<InvocationResponse, InvokeError>

Invokes a Lambda function. You can invoke a function synchronously (and wait for the response), or asynchronously. To invoke a function asynchronously, set InvocationType to Event.

For synchronous invocation, details about the function response, including errors, are included in the response body and headers. For either invocation type, you can find more information in the execution log and trace. To record function errors for asynchronous invocations, configure your function with a dead letter queue.

When an error occurs, your function may be invoked multiple times. Retry behavior varies by error type, client, event source, and invocation type. For example, if you invoke a function asynchronously and it returns an error, Lambda executes the function up to two more times. For more information, see Retry Behavior.

The status code in the API response doesn't reflect function errors. Error codes are reserved for errors that prevent your function from executing, such as permissions errors, limit errors, or issues with your function's code and configuration. For example, Lambda returns TooManyRequestsException if executing the function would cause you to exceed a concurrency limit at either the account level (ConcurrentInvocationLimitExceeded) or function level (ReservedFunctionConcurrentInvocationLimitExceeded).

For functions with a long timeout, your client might be disconnected during synchronous invocation while it waits for a response. Configure your HTTP client, SDK, firewall, proxy, or operating system to allow for long connections with timeout or keep-alive settings.

This operation requires permission for the lambda:InvokeFunction action.

fn invoke_async(
&self,
input: InvokeAsyncRequest
) -> RusotoFuture<InvokeAsyncResponse, InvokeAsyncError>

For asynchronous function invocation, use Invoke.

Invokes a function asynchronously.

fn list_aliases(
&self,
input: ListAliasesRequest
) -> RusotoFuture<ListAliasesResponse, ListAliasesError>

Returns a list of aliases for a Lambda function.

fn list_event_source_mappings(
&self,
input: ListEventSourceMappingsRequest
) -> RusotoFuture<ListEventSourceMappingsResponse, ListEventSourceMappingsError>

Lists event source mappings. Specify an EventSourceArn to only show event source mappings for a single event source.

fn list_functions(
&self,
input: ListFunctionsRequest
) -> RusotoFuture<ListFunctionsResponse, ListFunctionsError>

Returns a list of Lambda functions, with the version-specific configuration of each.

Set FunctionVersion to ALL to include all published versions of each function in addition to the unpublished version. To get more information about a function or version, use GetFunction.

fn list_layer_versions(
&self,
input: ListLayerVersionsRequest
) -> RusotoFuture<ListLayerVersionsResponse, ListLayerVersionsError>

Lists the versions of an AWS Lambda layer. Versions that have been deleted aren't listed. Specify a runtime identifier to list only versions that indicate that they're compatible with that runtime.

fn list_layers(
&self,
input: ListLayersRequest
) -> RusotoFuture<ListLayersResponse, ListLayersError>

Lists AWS Lambda layers and shows information about the latest version of each. Specify a runtime identifier to list only layers that indicate that they're compatible with that runtime.

fn list_tags(
&self,
input: ListTagsRequest
) -> RusotoFuture<ListTagsResponse, ListTagsError>

Returns a function's tags. You can also view tags with GetFunction.

fn list_versions_by_function(
&self,
input: ListVersionsByFunctionRequest
) -> RusotoFuture<ListVersionsByFunctionResponse, ListVersionsByFunctionError>

Returns a list of versions, with the version-specific configuration of each.

fn publish_layer_version(
&self,
input: PublishLayerVersionRequest
) -> RusotoFuture<PublishLayerVersionResponse, PublishLayerVersionError>

Creates an AWS Lambda layer from a ZIP archive. Each time you call PublishLayerVersion with the same version name, a new version is created.

Add layers to your function with CreateFunction or UpdateFunctionConfiguration.

fn publish_version(
&self,
input: PublishVersionRequest
) -> RusotoFuture<FunctionConfiguration, PublishVersionError>

Creates a version from the current code and configuration of a function. Use versions to create a snapshot of your function code and configuration that doesn't change.

AWS Lambda doesn't publish a version if the function's configuration and code haven't changed since the last version. Use UpdateFunctionCode or UpdateFunctionConfiguration to update the function before publishing a version.

Clients can invoke versions directly or with an alias. To create an alias, use CreateAlias.

fn put_function_concurrency(
&self,
input: PutFunctionConcurrencyRequest
) -> RusotoFuture<Concurrency, PutFunctionConcurrencyError>

Sets the maximum number of simultaneous executions for a function, and reserves capacity for that concurrency level.

Concurrency settings apply to the function as a whole, including all published versions and the unpublished version. Reserving concurrency both ensures that your function has capacity to process the specified number of events simultaneously, and prevents it from scaling beyond that level. Use GetFunction to see the current setting for a function.

Use GetAccountSettings to see your regional concurrency limit. You can reserve concurrency for as many functions as you like, as long as you leave at least 100 simultaneous executions unreserved for functions that aren't configured with a per-function limit. For more information, see Managing Concurrency.

fn remove_layer_version_permission(
&self,
input: RemoveLayerVersionPermissionRequest
) -> RusotoFuture<(), RemoveLayerVersionPermissionError>

Removes a statement from the permissions policy for a version of an AWS Lambda layer. For more information, see AddLayerVersionPermission.

fn remove_permission(
&self,
input: RemovePermissionRequest
) -> RusotoFuture<(), RemovePermissionError>

Revokes function-use permission from an AWS service or another account. You can get the ID of the statement from the output of GetPolicy.

fn tag_resource(
&self,
input: TagResourceRequest
) -> RusotoFuture<(), TagResourceError>

Adds tags to a function.

fn untag_resource(
&self,
input: UntagResourceRequest
) -> RusotoFuture<(), UntagResourceError>

Removes tags from a function.

fn update_alias(
&self,
input: UpdateAliasRequest
) -> RusotoFuture<AliasConfiguration, UpdateAliasError>

Updates the configuration of a Lambda function alias.

fn update_event_source_mapping(
&self,
input: UpdateEventSourceMappingRequest
) -> RusotoFuture<EventSourceMappingConfiguration, UpdateEventSourceMappingError>

Updates an event source mapping. You can change the function that AWS Lambda invokes, or pause invocation and resume later from the same location.

fn update_function_code(
&self,
input: UpdateFunctionCodeRequest
) -> RusotoFuture<FunctionConfiguration, UpdateFunctionCodeError>

Updates a Lambda function's code.

The function's code is locked when you publish a version. You can't modify the code of a published version, only the unpublished version.

fn update_function_configuration(
&self,
input: UpdateFunctionConfigurationRequest
) -> RusotoFuture<FunctionConfiguration, UpdateFunctionConfigurationError>

Modify the version-specific settings of a Lambda function.

These settings can vary between versions of a function and are locked when you publish a version. You can't modify the configuration of a published version, only the unpublished version.

To configure function concurrency, use PutFunctionConcurrency. To grant invoke permissions to an account or AWS service, use AddPermission.

Loading content...